Cite R packages used in a project

Find R packages used in a project, create a BibTeX file of citations, and generate a document with formatted package references, or a paragraph citing all packages used to be used directly within an Rmarkdown document (see examples).

cite_packages(
  output = c("file", "paragraph", "citekeys"),
  out.format = "html",
  citation.style = NULL,
  pkgs = "All",
  cite.tidyverse = FALSE,
  dependencies = FALSE,
  include.RStudio = FALSE,
  out.dir = getwd(),
  bib.file = "grateful-refs.bib",
  Rmd.file = "grateful-report.Rmd",
  out.name = "grateful-citations",
  ...
)

Arguments

output

Either "file" to generate a separate document with formatted citations for all packages; "paragraph" to return a paragraph with in-text citations of used packages, suitable to be used within an Rmarkdown document; or "citekeys" to return a vector with citation keys. In all cases, a BibTeX file with package references is saved on disk (see bib.file).

out.format

Output format when output = "file": either "html" (the default), docx" (Word), "pdf", or "md" (markdown). (Note that choosing "pdf" requires a working installation of LaTeX).

citation.style

Optional. Citation style to format references for a particular journal. See https://bookdown.org/yihui/rmarkdown-cookbook/bibliography.html.

pkgs

Character. Either "All" to include all packages used in scripts within the project/folder (the default), or "Session" to include only packages used in the current session. pkgs can also be a character vector of package names to get citations for (see examples).

cite.tidyverse

Logical. If TRUE, all tidyverse packages (dplyr, ggplot2, etc) will be collapsed into a single citation of the 'tidyverse'.

dependencies

Logical. Include the dependencies of your used packages? If TRUE, will include all the packages that your used packages depend on.

include.RStudio

Logical. If TRUE, adds a citation for the current version of RStudio.

out.dir

Directory to save the output document and a BibTeX file with the references. Default is the working directory.

bib.file

Desired name for the BibTeX file containing packages references ("grateful-refs.bib" by default).

Rmd.file

Desired name of the Rmarkdown document to be created if output = "file". Default is "grateful-report.Rmd".

out.name

Desired name of the output file containing the formatted references ("grateful-citations" by default).

...

Other parameters passed to dependencies.

Value

A file containing package references in BibTeX format, plus a file with formatted citations, or a paragraph with in-text citations of all packages, suitable to be used within Rmarkdown documents.

Details

cite_packages is a wrapper function that collects package names and versions and saves their citation information in a BibTeX file (using get_pkgs_info).

Then, the function is designed to handle three different use cases:

If output = "file", cite_packages() will generate an RMarkdown file which includes a paragraph with in-text citations of all packages, as well as a references list. This document can be knitted to various formats via out.format. References can be formatted for a particular journal using citation.style. Thus, output = "file" is best for obtaining a document separate from R, to just cut and paste citations.

If output = "paragraph", cite_packages() will return a paragraph with in-text citations of all packages, suitable to be used directly in an Rmarkdown document (see README). To do so, include a reference to the generated bib.file bibliography file in the YAML header of the Rmarkdown document.

Finally, you can use output = "citekeys" to obtain a vector of citation keys, and then call nocite_references within an Rmarkdown document to cite these packages in the reference list without mentioning them in the text.

Note

Before running grateful you might want to run stale_package_check on your scripts to check for unused packages before citing them.

Limitations

Citation keys are not guaranteed to be preserved when regenerated, particularly when packages are updated. This instability is not an issue when citations are used programmatically, as in the example below. But if references are put into the text manually, they may need to be updated periodically.

Examples