Get information about packages used in a project

This function scans the project for R packages used and returns a data frame with package names, version, and citation keys.

get_pkgs_info(
  pkgs = "All",
  cite.tidyverse = FALSE,
  dependencies = FALSE,
  out.dir = getwd(),
  bib.file = "grateful-refs.bib",
  include.RStudio = FALSE,
  ...
)

Arguments

pkgs

Character. Either "All" to include all packages used in scripts within the project/folder (the default), or "Session" to include only packages used in the current session. pkgs can also be a character vector of package names to get citations for (see examples).

cite.tidyverse

Logical. If TRUE, all tidyverse packages (dplyr, ggplot2, etc) will be collapsed into a single citation of the 'tidyverse'.

dependencies

Logical. Include the dependencies of your used packages? If TRUE, will include all the packages that your used packages depend on.

out.dir

Directory to save the BibTeX file with the references. Defaults to working directory.

bib.file

Name of the file to save the BibTeX references.

include.RStudio

Logical. If TRUE, adds a citation for the current version of RStudio.

...

Arguments passed on to scan_packages

Value

A data.frame with package info, and a file containing package references in BibTeX format.

Examples

if (FALSE) {
get_pkgs_info()
get_pkgs_info(pkgs = c("lme4", "mgcv"))
}